refactor: ensure bedrock code is client-side only refactor: moved animation logic into AnimatedDoor nitpick: improve PortalOffsets init code nitpick: ClientAdaptiveVariant SimpleExteriorModel -> ExteriorModel
