+++ src/main/java/dev/amble/ait/client/bedrock/BedrockAnimationTracker.java
@@ -1,13 +1,16 @@
-package dev.amble.ait.core.tardis.animation.v2.bedrock;
+package dev.amble.ait.client.bedrock;
 
 import dev.amble.ait.AITMod;
 import dev.drtheo.scheduler.api.TimeUnit;
 import dev.drtheo.scheduler.api.client.ClientScheduler;
 import dev.drtheo.scheduler.api.common.Scheduler;
 import dev.drtheo.scheduler.api.common.TaskStage;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 
 import java.util.function.Consumer;
 
+@Environment(EnvType.CLIENT)
 public class BedrockAnimationTracker {
 	public final BedrockAnimation animation;
 	private final Consumer<Runnable> ticker;
