+++ src/main/java/dev/amble/ait/client/models/consoles/BedrockConsoleModel.java
@@ -1,11 +1,9 @@
 package dev.amble.ait.client.models.consoles;
 
-import dev.amble.ait.api.tardis.link.v2.block.AbstractLinkableBlockEntity;
-import dev.amble.ait.client.models.AnimatedModel;
 import dev.amble.ait.client.tardis.ClientTardis;
 import dev.amble.ait.core.blockentities.ConsoleBlockEntity;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockAnimation;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockModel;
+import dev.amble.ait.client.bedrock.BedrockAnimation;
+import dev.amble.ait.client.bedrock.BedrockModel;
 import dev.amble.ait.core.tardis.handler.travel.TravelHandlerBase;
 import dev.amble.ait.data.datapack.DatapackConsole;
 import dev.amble.ait.data.schema.console.ConsoleVariantSchema;
@@ -13,7 +11,6 @@ import dev.amble.lib.api.Identifiable;
 import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.entity.AnimationState;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
