+++ src/main/java/dev/amble/ait/core/tardis/TardisManager.java
@@ -9,9 +9,6 @@ import com.google.gson.ExclusionStrategy;
 import com.google.gson.FieldAttributes;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockAnimation;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockAnimationAdapter;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockModel;
 import dev.amble.lib.data.DirectedBlockPos;
 import dev.amble.lib.data.DirectedGlobalPos;
 import org.jetbrains.annotations.Nullable;
