+++ src/main/java/dev/amble/ait/core/tardis/handler/travel/TravelHandler.java
@@ -131,7 +131,7 @@ public final class TravelHandler extends AnimatedTravelHandler implements Crasha
 
     @Override
     public boolean shouldTickAnimation() {
-        return true/* !this.waiting && this.getState().animated()*/;
+        return !this.waiting && this.getState().animated();
     }
 
     @Override
