+++ src/main/java/dev/amble/ait/data/schema/door/AnimatedDoor.java
@@ -1,6 +1,12 @@
 package dev.amble.ait.data.schema.door;
 
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockAnimationRegistry;
+import dev.amble.ait.client.bedrock.BedrockAnimationRegistry;
+import dev.amble.ait.client.tardis.ClientTardis;
+import dev.amble.ait.core.tardis.handler.DoorHandler;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.model.ModelPart;
+import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.util.math.Vec3d;
 
 import java.util.Optional;
@@ -20,5 +26,38 @@ public interface AnimatedDoor {
 
 	default Vec3d getOffset() {
 		return Vec3d.ZERO;
+	}
+
+	@Environment(EnvType.CLIENT)
+	default void runAnimations(ModelPart root, MatrixStack matrices, float tickDelta, ClientTardis tardis) {
+		DoorHandler doors = tardis.door();
+
+		Vec3d offset = this.getOffset().multiply(-1);
+		matrices.translate(offset.x, offset.y, offset.z);
+
+		Vec3d scale = this.getScale();
+		matrices.scale((float) scale.x, (float) scale.y, (float) scale.z);
+
+		matrices.push();
+		float leftProgress = doors.getLeftRot();
+		float rightProgress = doors.getRightRot();
+
+		float leftDelta;
+		if (leftProgress == 1 || leftProgress == 0) {
+			leftDelta = 0;
+		} else {
+			leftDelta = tickDelta;
+		}
+
+		float rightDelta;
+		if (rightProgress == 1 || rightProgress == 0) {
+			rightDelta = 0;
+		} else {
+			rightDelta = tickDelta;
+		}
+
+		this.getLeftAnimation().flatMap(BedrockAnimationRegistry.Reference::get).ifPresent(anim -> anim.apply(root, (int) (leftProgress * anim.animationLength * 20), leftDelta));
+		this.getRightAnimation().flatMap(BedrockAnimationRegistry.Reference::get).ifPresent(anim -> anim.apply(root, (int) (rightProgress * anim.animationLength * 20), rightDelta));
+		matrices.pop();
 	}
 }
