+++ src/main/java/dev/amble/ait/registry/impl/console/variant/ClientConsoleVariantRegistry.java
@@ -2,7 +2,7 @@ package dev.amble.ait.registry.impl.console.variant;
 
 import dev.amble.ait.client.models.consoles.BedrockConsoleModel;
 import dev.amble.ait.client.models.consoles.ConsoleModel;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.BedrockModelRegistry;
+import dev.amble.ait.client.bedrock.BedrockModelRegistry;
 import dev.amble.lib.register.datapack.DatapackRegistry;
 import org.joml.Vector3f;
 
@@ -11,7 +11,6 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 
 import dev.amble.ait.AITMod;
-import dev.amble.ait.client.models.consoles.SimpleConsoleModel;
 import dev.amble.ait.data.datapack.DatapackConsole;
 import dev.amble.ait.data.schema.console.ClientConsoleVariantSchema;
 import dev.amble.ait.data.schema.console.ConsoleVariantSchema;
