+++ src/main/java/dev/amble/ait/registry/impl/door/ClientDoorRegistry.java
@@ -1,11 +1,10 @@
 package dev.amble.ait.registry.impl.door;
 
 
-import dev.amble.ait.api.tardis.link.v2.block.AbstractLinkableBlockEntity;
 import dev.amble.ait.client.models.AnimatedModel;
 import dev.amble.ait.client.tardis.ClientTardis;
 import dev.amble.ait.core.blockentities.DoorBlockEntity;
-import dev.amble.ait.core.tardis.animation.v2.bedrock.*;
+import dev.amble.ait.client.bedrock.*;
 import dev.amble.ait.core.tardis.handler.DoorHandler;
 import dev.amble.ait.data.schema.door.AnimatedDoor;
 import dev.amble.ait.data.schema.door.DatapackDoor;
@@ -132,35 +131,10 @@ public class ClientDoorRegistry extends DatapackRegistry<ClientDoorSchema> {
 	                public void renderWithAnimations(ClientTardis tardis, DoorBlockEntity be, ModelPart root, MatrixStack matrices, VertexConsumer vertices, int light, int overlay, float red, float green, float blue, float pAlpha, float tickDelta) {
 		                matrices.push();
 
-		                DoorHandler doors = tardis.door();
 		                DoorSchema schema = tardis.getExterior().getVariant().door();
 
 		                if (schema instanceof AnimatedDoor animDoor) {
-			                Vec3d offset = animDoor.getOffset().multiply(-1);
-			                matrices.translate(offset.x, offset.y, offset.z);
-
-			                Vec3d scale = animDoor.getScale();
-			                matrices.scale((float) scale.x, (float) scale.y, (float) scale.z);
-
-			                float leftProgress = doors.getLeftRot();
-			                float rightProgress = doors.getRightRot();
-
-			                float leftDelta;
-			                if (leftProgress == 1 || leftProgress == 0) {
-				                leftDelta = 0;
-			                } else {
-				                leftDelta = tickDelta;
-			                }
-
-			                float rightDelta;
-			                if (rightProgress == 1 || rightProgress == 0) {
-				                rightDelta = 0;
-			                } else {
-				                rightDelta = tickDelta;
-			                }
-
-			                animDoor.getLeftAnimation().flatMap(BedrockAnimationRegistry.Reference::get).ifPresent(anim -> anim.apply(root, (int) (leftProgress * anim.animationLength * 20), leftDelta));
-			                animDoor.getRightAnimation().flatMap(BedrockAnimationRegistry.Reference::get).ifPresent(anim -> anim.apply(root, (int) (rightProgress * anim.animationLength * 20), rightDelta));
+							animDoor.runAnimations(root, matrices, tickDelta, tardis);
 		                }
 		                root.render(matrices, vertices, light, overlay, red, green, blue, pAlpha);
